feature_extraction/requirements.txt
# torch
# torchvision
h5py
numpy
tqdm
einops
timm
xformers
webdataset
open_clip_torch
opencv-python
transformers
